feat: separate brillig and brillig_vm
