+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 
-members = ["acir_field", "acir", "acvm", "stdlib", "brillig_vm"]
+members = ["acir_field", "acir", "acvm", "stdlib", "brillig", "brillig_vm"]
 
 [workspace.package]
 authors = ["The Noir Team <team@noir-lang.org>"]
@@ -13,6 +13,7 @@ repository = "https://github.com/noir-lang/acvm/"
 acir = { version = "0.17.0", path = "acir", default-features = false }
 acir_field = { version = "0.17.0", path = "acir_field", default-features = false }
 stdlib = { package = "acvm_stdlib", version = "0.17.0", path = "stdlib", default-features = false }
+brillig = { version = "0.17.0", path = "brillig", default-features = false }
 rmp-serde = "1.1.0"
 
 num-bigint = "0.4"
