+++ acvm/Cargo.toml
@@ -18,6 +18,7 @@ rmp-serde.workspace = true
 
 acir.workspace = true
 stdlib.workspace = true
+brillig_vm = { version = "0.17.0", path = "../brillig_vm", default-features = false }
 
 blake2.workspace = true
 sha2.workspace = true
@@ -30,8 +31,8 @@ async-trait = "0.1"
 
 [features]
 default = ["bn254"]
-bn254 = ["acir/bn254", "stdlib/bn254"]
-bls12_381 = ["acir/bls12_381", "stdlib/bls12_381"]
+bn254 = ["acir/bn254", "stdlib/bn254", "brillig_vm/bn254"]
+bls12_381 = ["acir/bls12_381", "stdlib/bls12_381", "brillig_vm/bls12_381"]
 
 [dev-dependencies]
 rand = "0.8.5"
