+++ brillig_vm/src/registers.rs
@@ -1,5 +1,4 @@
-use crate::Value;
-use serde::{Deserialize, Serialize};
+use brillig::{RegisterIndex, Value};
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct Registers {
@@ -12,22 +11,6 @@ pub struct Registers {
 /// As well, catches obvious erroneous use of registers.
 /// This can be revisited if it proves not enough.
 const MAX_REGISTERS: usize = 2_usize.pow(16);
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RegisterIndex(usize);
-
-/// `RegisterIndex` refers to the index in VM register space.
-impl RegisterIndex {
-    pub fn to_usize(self) -> usize {
-        self.0
-    }
-}
-
-impl From<usize> for RegisterIndex {
-    fn from(value: usize) -> Self {
-        RegisterIndex(value)
-    }
-}
 
 /// Registers will store field element values during the
 /// duration of the execution of the bytecode.
